backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic[email]==2.5.3
python-multipart==0.0.6
httpx==0.26.0
bcrypt==4.1.2

# IA e RAG
langchain==0.1.0
langchain-community==0.0.10
qdrant-client==1.7.0
openai-whisper==20231117
pydub==0.25.1
gTTS==2.5.0

# Processamento de documentos
pypdf==3.17.4
python-docx==1.1.0
chardet==5.2.0

# Ollama
requests==2.31.0
